==> nbs_arxiv/01-experimental-modelling.ipynb <==
%load_ext autoreload
%autoreload 2

import pandas as pd
from tsm.data_utils import get_data_sample, time_processing, df_to_x_y, train_dev_test_split

import numpy as np

from tsm.model_baselines import dummy_regressor_accuracy
from tsm.eval_metrics  import root_mean_squared_log_error

from tsm.data_utils import ordinal2wave

## Load and Prepare Data for Experiments

train_data = pd.read_pickle('kaggle/input/ashrae-energy-prediction/train.pkl')

### Full Training Data

trdf, dvdf, tsdf = train_dev_test_split(train_data)

x_idx = [0, 1] + [x for x in range(4, len(list(train_data)))]
y_idx = 3

tr_x, tr_y = df_to_x_y(trdf, x_idx, y_idx)
dv_x, dv_y = df_to_x_y(dvdf, x_idx, y_idx)
ts_x, ts_y = df_to_x_y(tsdf, x_idx, y_idx)

# Dummy regressor
dummy = dummy_regressor_accuracy(tr_x, tr_y, root_mean_squared_log_error)

# Smarter Dummy: takes the mean by meter type
def smart_dummy_predict(tr_df, tr_x, meter_idx_in_df: int = 1):
    mean_by_type = tr_df.groupby('meter')['meter_reading'].mean().to_dict()
    return np.array([mean_by_type[r] for r in tr_x[:, meter_idx_in_df]])

# Smart dummy training accuracy
root_mean_squared_log_error(smart_dummy_predict(trdf, tr_x), tr_y)

# Smart dummy test accuracy
root_mean_squared_log_error(smart_dummy_predict(trdf, ts_x), ts_y)

## Let me experiment with a single site before scaling

train_data.building_id.value_counts()

sample_site = train_data[train_data.building_id == 1241]

len(sample_site)

ordinal2wave('dt_m', sample_site)
ordinal2wave('dt_m', sample_site)
ordinal2wave('dt_w', sample_site)
ordinal2wave('dt_d', sample_site)
ordinal2wave('dt_hour', sample_site)
ordinal2wave('dt_day_week', sample_site)
ordinal2wave('dt_day_month', sample_site)
ordinal2wave('dt_week_month', sample_site)

tr_smp, ts_smp = train_dev_test_split(sample_site, train_pct=0.75, dev_pct=0)

x_idx = [0, 1] + [x for x in range(4, len(list(sample_site)))]
y_idx = 3

tr_smpx, tr_smpy = df_to_x_y(tr_smp, x_idx, y_idx)
ts_smpx, ts_smpy = df_to_x_y(ts_smp, x_idx, y_idx)

# Smart dummy training accuracy
root_mean_squared_log_error(smart_dummy_predict(tr_smp, tr_smpx), tr_smpy)

# Smart dummy training accuracy
root_mean_squared_log_error(smart_dummy_predict(tr_smp, ts_smpx), ts_smpy)

# Let's try a Random Forest regression
from sklearn.ensemble import RandomForestRegressor
rfrg = RandomForestRegressor(n_jobs=-1, max_depth=5, n_estimators=39)

rfrg.fit(tr_smpx, tr_smpy)

root_mean_squared_log_error(rfrg.predict(tr_smpx), tr_smpy)

root_mean_squared_log_error(rfrg.predict(ts_smpx), ts_smpy)

pd.set_option('display.max_columns', 200)
sample_site.describe()

## Time Series Modelling

### Let's start using no feature other than meter reading and meter type

from datetime import timedelta

sample_site_ts = sample_site[['meter', 'meter_reading', 'timestamp']]

from tsm.data_utils import add_k_past_lags

sample_site_ts = sample_site_ts.progress_apply(lambda r: add_k_past_lags(r, sample_site_ts, 'timestamp', 'meter_reading', 'meter'), axis=1)

sample_site_ts.dropna(inplace=True)

sample_site_ts

tr_smpts, ts_smpts = train_dev_test_split(sample_site_ts, train_pct=0.75, dev_pct=0)

tsx_idx = [0] + [x for x in range(3, len(list(sample_site_ts)))]
tsy_idx = 1

tr_smpx, tr_smpy = df_to_x_y(tr_smpts, tsx_idx, tsy_idx)
ts_smpx, ts_smpy = df_to_x_y(ts_smpts, tsx_idx, tsy_idx)

rfrg = RandomForestRegressor(n_jobs=-1, max_depth=25, n_estimators=39)
rfrg.fit(tr_smpx, tr_smpy)

root_mean_squared_log_error(rfrg.predict(tr_smpx), tr_smpy)

root_mean_squared_log_error(rfrg.predict(ts_smpx), ts_smpy)

# 10 past lags + features?
sample_ts_with_feats = sample_site_ts.merge(sample_site)

tr_smptsf, ts_smptsf = train_dev_test_split(sample_ts_with_feats, train_pct=0.75, dev_pct=0)

sample_ts_with_feats.head()

tsfx_idx = [0] + [x for x in range(3, len(list(sample_ts_with_feats)))]
tsfy_idx = 1

tr_smfpx, tr_smfpy = df_to_x_y(tr_smptsf, tsfx_idx, tsfy_idx)
ts_smfpx, ts_smfpy = df_to_x_y(ts_smptsf, tsfx_idx, tsfy_idx)

rfrg = RandomForestRegressor(n_jobs=-1, max_depth=5, n_estimators=39)
rfrg.fit(tr_smfpx, tr_smfpy)

root_mean_squared_log_error(rfrg.predict(tr_smfpx), tr_smfpy)

root_mean_squared_log_error(rfrg.predict(ts_smfpx), ts_smfpy)

tr_smptsf.shape

# Save processed features before leaving
sample_ts_with_feats.to_pickle('kaggle/input/ashrae-energy-prediction/store_1241_10lags_train.pkl')
