==> nbs_arxiv/02-experiments-level3-0s.ipynb <==
%load_ext autoreload
%autoreload 2

import pandas as pd
from tsm.data_selector import data_subset_by_dict
import matplotlib.pyplot as plt
import numpy as np

train_data = pd.read_pickle('kaggle/input/ashrae-energy-prediction/train.pkl')

## Level 3A)
* 1 site - X buildings
* Meter reading: 0
* All features
--> Studying the missing values

level_3a = {'site_id': 9, 'meter': 0}
level_3a_cols = list(train_data)

data_3a = data_subset_by_dict(train_data, level_3a, level_3a_cols)

data_3a.head()

## Baseline 1) LGBM on hold out buildings

buildings = list(set(data_3a.building_id))

test_builds = [x for ix, x in enumerate(buildings) if ix % 5 == 0]

round(len(test_builds) / len(buildings) * 100, 3), '%'

train_builds = [x for x in buildings if x not in test_builds]

data_3a_tr = data_3a[data_3a.building_id.isin(train_builds)]
data_3a_val = data_3a[data_3a.building_id.isin(test_builds[:int(len(test_builds)*0.25)])]
data_3a_ts = data_3a[data_3a.building_id.isin(test_builds[int(len(test_builds)*0.25):])]

import lightgbm as lgb

from tsm.eval_metrics import root_mean_squared_log_error

feat_cols = [x for x in list(train) if x not in ['meter_reading', 'timestamp']]

data_3a_tr_x = data_3a_tr[feat_cols]
data_3a_tr_y = data_3a_tr.meter_reading
data_3a_val_x = data_3a_val[feat_cols]
data_3a_val_y = data_3a_val.meter_reading
data_3a_ts_x = data_3a_ts[feat_cols]
data_3a_ts_y = data_3a_ts.meter_reading

SEED = 7

lgb_reg_params = {'objective':'regression',  'boosting_type':'gbdt', 'metric':'rmse',
                  'n_jobs':-1, 'learning_rate':0.07, 'num_leaves': 2**8, 'max_depth':-1,
                  'tree_learner':'serial', 'colsample_bytree': 0.7, 'subsample_freq':1,
                  'subsample':0.5, 'max_bin':255, 'verbose':1, 'seed': SEED, 'early_stopping_rounds':200} 
lgb_reg_params_2 = {
            'boosting_type': 'gbdt',
            'objective': 'regression',
            'metric': {'rmse'},
            'learning_rate': 0.5,
            'feature_fraction': 0.8,
            'bagging_fraction': 0.8,
            'bagging_freq' : 5
            }

#### Base

lgb_train = lgb.Dataset(data_3a_tr_x, data_3a_tr_y)
lgb_eval = lgb.Dataset(data_3a_val_x, data_3a_val_y)
gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                        early_stopping_rounds=20, verbose_eval = 20)

lgbm_hat = (gbm_regress.predict(data_3a_ts_x, num_iteration=gbm_regress.best_iteration))

root_mean_squared_log_error(lgbm_hat, data_3a_ts_y)

train = data_3a

#### V1 - Simply take log of target

data_3a1_tr_y = np.log(data_3a_tr_y)
data_3a1_val_y = np.log(data_3a_val_y)

data_3a1_tr_x = data_3a_tr_x[np.isfinite(data_3a1_tr_y)]
data_3a1_tr_y = data_3a1_tr_y[np.isfinite(data_3a1_tr_y)]
data_3a1_val_x = data_3a_val_x[np.isfinite(data_3a1_val_y)]
data_3a1_val_y = data_3a1_val_y[np.isfinite(data_3a1_val_y)]

lgb_train = lgb.Dataset(data_3a1_tr_x, data_3a1_tr_y)
lgb_eval = lgb.Dataset(data_3a1_val_x, data_3a1_val_y)
gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                        early_stopping_rounds=20, verbose_eval = 20)

lgbm_hat = np.exp(gbm_regress.predict(data_3a_ts_x, num_iteration=gbm_regress.best_iteration))

root_mean_squared_log_error(lgbm_hat, data_3a_ts_y)

#### V2 - two steps model (0s and values)

data_3a2_tr_y = np.log(data_3a_tr_y)
data_3a2_val_y = np.log(data_3a_val_y)

lgb_train = lgb.Dataset(data_3a_tr_x, data_3a2_tr_y > 0)
lgb_eval = lgb.Dataset(data_3a_val_x, data_3a2_val_y > 0)
class_params = {'boosting_type': 'gbdt', 'objective': 'binary', 'metric': {'binary_logloss'},
                'learning_rate': 0.1, 'feature_fraction': 0.8, 'bagging_fraction': 0.8,
                'bagging_freq' : 5}
gbm_class = lgb.train(class_params,
            lgb_train,
            num_boost_round=2000,
            valid_sets=(lgb_train, lgb_eval),
            early_stopping_rounds=20,
            verbose_eval = 20)
lgb_train = lgb.Dataset(data_3a_tr_x[data_3a2_tr_y > 0], data_3a2_tr_y[data_3a2_tr_y > 0])
lgb_eval = lgb.Dataset(data_3a_val_x[data_3a2_val_y > 0] , data_3a2_val_y[data_3a2_val_y > 0])
gbm_regress = lgb.train(lgb_reg_params,
            lgb_train,
            num_boost_round=2000,
            valid_sets=(lgb_train, lgb_eval),
            early_stopping_rounds=20,
            verbose_eval = 20)

lgbm_hat = (gbm_class.predict(data_3a_ts_x, num_iteration=gbm_class.best_iteration) > .5) *\
    (gbm_regress.predict(data_3a_ts_x, num_iteration=gbm_regress.best_iteration))

root_mean_squared_log_error(np.exp(lgbm_hat), data_3a_ts_y)

### Base No 0s

lgb_train = lgb.Dataset(data_3a_tr_x[data_3a_tr_y > 0], data_3a_tr_y[data_3a_tr_y > 0])
lgb_eval = lgb.Dataset(data_3a_val_x[data_3a_val_y > 0], data_3a_val_y[data_3a_val_y > 0])
gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                        early_stopping_rounds=20, verbose_eval = 20)

lgbm_hat = (gbm_regress.predict(data_3a_ts_x, num_iteration=gbm_regress.best_iteration))

root_mean_squared_log_error(lgbm_hat, data_3a_ts_y)

## Baseline 2) Random Forest on 5-fold cross validation

from sklearn.model_selection import KFold

from sklearn.ensemble import RandomForestRegressor, RandomForestClassifier

rfr = RandomForestRegressor(max_depth=10, n_jobs=-1, n_estimators=11)

K = 5
kf = KFold(n_splits=K)

nrmlses = []
for tr_ix, ts_ix in kf.split(data_3a):
    data_3a2_tr_x = data_3a.iloc[tr_ix,:][feat_cols]
    data_3a2_tr_y = data_3a.iloc[tr_ix,:].meter_reading
    rfr.fit(data_3a2_tr_x, data_3a2_tr_y)
    data_3a2_ts_x = data_3a.iloc[ts_ix,:][feat_cols]
    data_3a2_ts_y = data_3a.iloc[ts_ix,:].meter_reading
    nrmlse = root_mean_squared_log_error(rfr.predict(data_3a2_ts_x), data_3a2_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)

sum(nrmlses) / K

nrmlses = []
for tr_ix, ts_ix in kf.split(data_3a):
    data_3a2_tr_x = data_3a.iloc[tr_ix[:int(len(tr_ix)*.9)],:][feat_cols]
    data_3a2_tr_y = data_3a.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading
    data_3a2_val_x = data_3a.iloc[tr_ix[int(len(tr_ix)*.9):],:][feat_cols]
    data_3a2_val_y = data_3a.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading
    lgb_train = lgb.Dataset(data_3a2_tr_x, data_3a2_tr_y)
    lgb_eval = lgb.Dataset(data_3a2_val_x, data_3a2_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    data_3a2_ts_x = data_3a.iloc[ts_ix,:][feat_cols]
    data_3a2_ts_y = data_3a.iloc[ts_ix,:].meter_reading
    lgbm_hat = (gbm_regress.predict(data_3a2_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(lgbm_hat, data_3a2_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)

print('FinalScore:', sum(nrmlses) / K)

nrmlses = []
for tr_ix, ts_ix in kf.split(data_3a):
    data_3a2_tr_x = data_3a.iloc[tr_ix[:int(len(tr_ix)*.9)],:][feat_cols]
    data_3a2_tr_y = np.log1p(data_3a.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading)
    data_3a2_val_x = data_3a.iloc[tr_ix[int(len(tr_ix)*.9):],:][feat_cols]
    data_3a2_val_y = np.log1p(data_3a.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading)
    lgb_train = lgb.Dataset(data_3a2_tr_x, data_3a2_tr_y)
    lgb_eval = lgb.Dataset(data_3a2_val_x, data_3a2_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    data_3a2_ts_x = data_3a.iloc[ts_ix,:][feat_cols]
    data_3a2_ts_y = data_3a.iloc[ts_ix,:].meter_reading
    lgbm_hat = (gbm_regress.predict(data_3a2_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(np.exp(lgbm_hat), data_3a2_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)
print('FinalScore:', sum(nrmlses) / K)

# LogY No0s
nrmlses = []
for tr_ix, ts_ix in kf.split(data_3a):
    data_3a2_tr_x = data_3a.iloc[tr_ix[:int(len(tr_ix)*.9)],:][feat_cols]
    data_3a2_tr_y = np.log1p(data_3a.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading)
    data_3a2_val_x = data_3a.iloc[tr_ix[int(len(tr_ix)*.9):],:][feat_cols]
    data_3a2_val_y = np.log1p(data_3a.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading)
    data_3a2_tr_x = data_3a2_tr_x[data_3a2_tr_y > 0]
    data_3a2_tr_y = data_3a2_tr_y[data_3a2_tr_y > 0]
    data_3a2_val_x = data_3a2_val_x[data_3a2_val_y > 0]
    data_3a2_val_y = data_3a2_val_y[data_3a2_val_y > 0]
    lgb_train = lgb.Dataset(data_3a2_tr_x, data_3a2_tr_y)
    lgb_eval = lgb.Dataset(data_3a2_val_x, data_3a2_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    data_3a2_ts_x = data_3a.iloc[ts_ix,:][feat_cols]
    data_3a2_ts_y = data_3a.iloc[ts_ix,:].meter_reading
    lgbm_hat = (gbm_regress.predict(data_3a2_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(np.exp(lgbm_hat), data_3a2_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)
print('FinalScore:', sum(nrmlses) / K)

# 2-steps
nrmlses = []
for tr_ix, ts_ix in kf.split(data_3a):
    data_3a2_tr_x = data_3a.iloc[tr_ix[:int(len(tr_ix)*.9)],:][feat_cols]
    data_3a2_tr_y = np.log1p(data_3a.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading)
    data_3a2_val_x = data_3a.iloc[tr_ix[int(len(tr_ix)*.9):],:][feat_cols]
    data_3a2_val_y = np.log1p(data_3a.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading)
    lgb_train = lgb.Dataset(data_3a2_tr_x, data_3a2_tr_y > 0)
    lgb_eval = lgb.Dataset(data_3a2_val_x, data_3a2_val_y > 0)
    class_params = {'boosting_type': 'gbdt', 'objective': 'binary', 'metric': {'binary_logloss'},
                    'learning_rate': 0.1, 'feature_fraction': 0.8, 'bagging_fraction': 0.8,
                    'bagging_freq' : 5}
    gbm_class = lgb.train(class_params,
                lgb_train,
                num_boost_round=2000,
                valid_sets=(lgb_train, lgb_eval),
                early_stopping_rounds=20,
                verbose_eval = 20)
    lgb_train = lgb.Dataset(data_3a2_tr_x[data_3a2_tr_y > 0], data_3a2_tr_y[data_3a2_tr_y > 0])
    lgb_eval = lgb.Dataset(data_3a2_val_x[data_3a2_val_y > 0] , data_3a2_val_y[data_3a2_val_y > 0])
    gbm_regress = lgb.train(lgb_reg_params,
                lgb_train,
                num_boost_round=2000,
                valid_sets=(lgb_train, lgb_eval),
                early_stopping_rounds=20,
                verbose_eval = 20)
    data_3a2_ts_x = data_3a.iloc[ts_ix,:][feat_cols]
    data_3a2_ts_y = data_3a.iloc[ts_ix,:].meter_reading
    lgbm_hat = (gbm_class.predict(data_3a2_ts_x, num_iteration=gbm_class.best_iteration) > .5) *\
    (gbm_regress.predict(data_3a2_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(np.exp(lgbm_hat), data_3a2_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)
print('FinalScore:', sum(nrmlses) / K)

nrmlses = []
for tr_ix, ts_ix in kf.split(data_3a):
    data_3a2_tr_x = data_3a.iloc[tr_ix[:int(len(tr_ix)*.9)],:][feat_cols]
    data_3a2_tr_y = data_3a.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading
    data_3a2_val_x = data_3a.iloc[tr_ix[int(len(tr_ix)*.9):],:][feat_cols]
    data_3a2_val_y = data_3a.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading
    data_3a2_tr_x = data_3a2_tr_x[data_3a2_tr_y > 0]
    data_3a2_tr_y = data_3a2_tr_y[data_3a2_tr_y > 0]
    data_3a2_val_x = data_3a2_val_x[data_3a2_val_y > 0]
    data_3a2_val_y = data_3a2_val_y[data_3a2_val_y > 0]
    lgb_train = lgb.Dataset(data_3a2_tr_x, data_3a2_tr_y)
    lgb_eval = lgb.Dataset(data_3a2_val_x, data_3a2_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    data_3a2_ts_x = data_3a.iloc[ts_ix,:][feat_cols]
    data_3a2_ts_y = data_3a.iloc[ts_ix,:].meter_reading
    lgbm_hat = (gbm_regress.predict(data_3a2_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(lgbm_hat, data_3a2_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)

sum(nrmlses) / K
