==> nbs_arxiv/02-experiments-level4-buildingID.ipynb <==
.. I am thinking to not use building_id but instead:
- mean meter reading by meter type
- std of meter reading by meter type
- mean meter reading by meter type by week of year
- std meter reading by meter type by week of year

%load_ext autoreload
%autoreload 2

import pandas as pd
from tsm.data_selector import data_subset_by_dict
import matplotlib.pyplot as plt
import numpy as np

train_data = pd.read_pickle('kaggle/input/ashrae-energy-prediction/train.pkl')

## Level 4A)
* 1 site - 124 buildings
* 1 meter reading - meter 0
* Building ID

level_4a = {'site_id': 14, 'meter': 0}
level_4a_cols = ['building_id', 'meter_reading', 'dt_m', 'dt_w', 'dt_d', 'dt_hour', 
                 'dt_day_week', 'dt_day_month', 'dt_week_month']

data_4a = data_subset_by_dict(train_data, level_4a, level_4a_cols)

data_4a.head()

import lightgbm as lgb

from tsm.eval_metrics import root_mean_squared_log_error

SEED = 7

lgb_reg_params = {'objective':'regression',  'boosting_type':'gbdt', 'metric':'rmse',
                  'n_jobs':-1, 'learning_rate':0.07, 'num_leaves': 2**8, 'max_depth':-1,
                  'tree_learner':'serial', 'colsample_bytree': 0.7, 'subsample_freq':1,
                  'subsample':0.5, 'max_bin':255, 'verbose':1, 'seed': SEED, 'early_stopping_rounds': 40} 

# Cross Validation
from sklearn.model_selection import KFold

K = 5
kf = KFold(n_splits=K)

feat_cols = ['building_id', 'dt_m', 'dt_w', 'dt_d', 'dt_hour', 
             'dt_day_week', 'dt_day_month', 'dt_week_month']

nrmlses = []
for tr_ix, ts_ix in kf.split(data_4a):
    data_4a1_tr_x = data_4a.iloc[tr_ix[:int(len(tr_ix)*.9)],:][feat_cols]
    data_4a1_tr_y = data_4a.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading
    data_4a1_val_x = data_4a.iloc[tr_ix[int(len(tr_ix)*.9):],:][feat_cols]
    data_4a1_val_y = data_4a.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading
    lgb_train = lgb.Dataset(data_4a1_tr_x, data_4a1_tr_y)
    lgb_eval = lgb.Dataset(data_4a1_val_x, data_4a1_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    data_4a1_ts_x = data_4a.iloc[ts_ix,:][feat_cols]
    data_4a1_ts_y = data_4a.iloc[ts_ix,:].meter_reading
    lgbm_hat = (gbm_regress.predict(data_4a1_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(lgbm_hat, data_4a1_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)
print('FinalScore:', sum(nrmlses) / K)

feat_cols = ['building_id_mean', 'dt_m', 'dt_w', 'dt_d', 'dt_hour', 
             'dt_day_week', 'dt_day_month', 'dt_week_month']

# Simple building_id mean
nrmlses = []
for tr_ix, ts_ix in kf.split(data_4a):
    full_training_data = data_4a.iloc[tr_ix[:int(len(tr_ix)*.9)],:]
    building_id_means = full_training_data.groupby('building_id')['meter_reading'].mean().to_dict()
    full_training_data['building_id_mean'] = full_training_data.building_id.apply(lambda x: building_id_means[x])
    data_4a2_tr_x = full_training_data[feat_cols]
    data_4a2_tr_y = data_4a.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading
    data_4a2_val_x = data_4a.iloc[tr_ix[int(len(tr_ix)*.9):],:]
    data_4a2_val_x['building_id_mean'] = data_4a2_val_x.building_id.apply(lambda x: building_id_means[x])
    data_4a2_val_x = data_4a2_val_x[feat_cols]
    data_4a2_val_y = data_4a.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading
    lgb_train = lgb.Dataset(data_4a2_tr_x, data_4a2_tr_y)
    lgb_eval = lgb.Dataset(data_4a2_val_x, data_4a2_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    data_4a2_ts_x = data_4a.iloc[ts_ix,:]
    data_4a2_ts_x['building_id_mean'] = data_4a2_ts_x.building_id.apply(lambda x: building_id_means[x])
    data_4a2_ts_x = data_4a2_ts_x[feat_cols]
    data_4a2_ts_y = data_4a.iloc[ts_ix,:].meter_reading
    lgbm_hat = (gbm_regress.predict(data_4a2_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(lgbm_hat, data_4a2_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)
print('FinalScore:', sum(nrmlses) / K)

# Building_id mean by month
nrmlses = []
for tr_ix, ts_ix in kf.split(data_4a):
    # Building Means
    full_training_data = data_4a.iloc[tr_ix[:int(len(tr_ix)*.9)],:]
    building_id_means = full_training_data.groupby(['building_id', 'dt_m'], as_index=False)['meter_reading'].mean()
    bid_dicts = {}
    for bid in list(set(building_id_means.building_id)):
        bid_data = building_id_means[building_id_means.building_id == bid][['dt_m', 'meter_reading']].to_dict('records')
        bid_dicts[bid] = bid_data
    # Using Means
    bid_means_tr = pd.DataFrame(full_training_data.building_id.apply(lambda x: [x['meter_reading'] for x in bid_dicts[x]]).values.tolist(),
                      columns=['building_id_dt_m_{}_mean'.format(m) for m in [mm['dt_m'] for mm in bid_dicts[bid]]]).reset_index(drop=True)
    data_4a2_tr_x = full_training_data.reset_index(drop=True).merge(bid_means_tr, left_index=True, right_index=True)
    data_4a2_tr_x = data_4a2_tr_x[[x for x in list(data_4a2_tr_x) if x not in ['meter_reading', 'building_id']]].fillna(0)
    data_4a2_tr_y = data_4a.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading
    
    data_4a2_val_x = data_4a.iloc[tr_ix[int(len(tr_ix)*.9):],:]
    bid_means_val = pd.DataFrame(data_4a2_val_x.building_id.apply(lambda x: [x['meter_reading'] for x in bid_dicts[x]]).values.tolist(),
                  columns=['building_id_dt_m_{}_mean'.format(m) for m in [mm['dt_m'] for mm in bid_dicts[bid]]]).reset_index(drop=True)
    data_4a2_val_x = data_4a2_val_x.reset_index(drop=True).merge(bid_means_val, left_index=True, right_index=True)
    data_4a2_val_x = data_4a2_val_x[[x for x in list(data_4a2_val_x) if x not in ['meter_reading', 'building_id']]].fillna(0)
    data_4a2_val_y = data_4a.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading
    
    lgb_train = lgb.Dataset(data_4a2_tr_x, data_4a2_tr_y)
    lgb_eval = lgb.Dataset(data_4a2_val_x, data_4a2_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    
    data_4a2_ts_x = data_4a.iloc[ts_ix,:]
    bid_means_ts = pd.DataFrame(data_4a2_ts_x.building_id.apply(lambda x: [x['meter_reading'] for x in bid_dicts[x]]).values.tolist(),
                  columns=['building_id_dt_m_{}_mean'.format(m) for m in [mm['dt_m'] for mm in bid_dicts[bid]]]).reset_index(drop=True)
    data_4a2_ts_x = data_4a2_ts_x.reset_index(drop=True).merge(bid_means_ts, left_index=True, right_index=True)
    data_4a2_ts_x = data_4a2_ts_x[[x for x in list(data_4a2_ts_x) if x not in ['meter_reading', 'building_id']]].fillna(0)
    data_4a2_ts_y = data_4a.iloc[ts_ix,:].meter_reading
    
    lgbm_hat = (gbm_regress.predict(data_4a2_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(lgbm_hat, data_4a2_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)
print('FinalScore:', sum(nrmlses) / K)

# Building_id mean by week of the year
nrmlses = []
for tr_ix, ts_ix in kf.split(data_4a):
    # Building Means
    full_training_data = data_4a.iloc[tr_ix[:int(len(tr_ix)*.9)],:]
    building_id_means = full_training_data.groupby(['building_id', 'dt_w'], as_index=False)['meter_reading'].mean()
    bid_dicts = {}
    for bid in list(set(building_id_means.building_id)):
        bid_data = building_id_means[building_id_means.building_id == bid][['dt_w', 'meter_reading']].to_dict('records')
        bid_dicts[bid] = bid_data
    # Using Means
    bid_means_tr = pd.DataFrame(full_training_data.building_id.apply(lambda x: [x['meter_reading'] for x in bid_dicts[x]]).values.tolist(),
                      columns=['building_id_dt_w_{}_mean'.format(m) for m in [mm['dt_w'] for mm in bid_dicts[bid]]]).reset_index(drop=True)
    data_4a2_tr_x = full_training_data.reset_index(drop=True).merge(bid_means_tr, left_index=True, right_index=True)
    data_4a2_tr_x = data_4a2_tr_x[[x for x in list(data_4a2_tr_x) if x not in ['meter_reading', 'building_id']]].fillna(0)
    data_4a2_tr_y = data_4a.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading
    
    data_4a2_val_x = data_4a.iloc[tr_ix[int(len(tr_ix)*.9):],:]
    bid_means_val = pd.DataFrame(data_4a2_val_x.building_id.apply(lambda x: [x['meter_reading'] for x in bid_dicts[x]]).values.tolist(),
                  columns=['building_id_dt_w_{}_mean'.format(m) for m in [mm['dt_w'] for mm in bid_dicts[bid]]]).reset_index(drop=True)
    data_4a2_val_x = data_4a2_val_x.reset_index(drop=True).merge(bid_means_val, left_index=True, right_index=True)
    data_4a2_val_x = data_4a2_val_x[[x for x in list(data_4a2_val_x) if x not in ['meter_reading', 'building_id']]].fillna(0)
    data_4a2_val_y = data_4a.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading
    
    lgb_train = lgb.Dataset(data_4a2_tr_x, data_4a2_tr_y)
    lgb_eval = lgb.Dataset(data_4a2_val_x, data_4a2_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    
    data_4a2_ts_x = data_4a.iloc[ts_ix,:]
    bid_means_ts = pd.DataFrame(data_4a2_ts_x.building_id.apply(lambda x: [x['meter_reading'] for x in bid_dicts[x]]).values.tolist(),
                  columns=['building_id_dt_w_{}_mean'.format(m) for m in [mm['dt_w'] for mm in bid_dicts[bid]]]).reset_index(drop=True)
    data_4a2_ts_x = data_4a2_ts_x.reset_index(drop=True).merge(bid_means_ts, left_index=True, right_index=True)
    data_4a2_ts_x = data_4a2_ts_x[[x for x in list(data_4a2_ts_x) if x not in ['meter_reading', 'building_id']]].fillna(0)
    data_4a2_ts_y = data_4a.iloc[ts_ix,:].meter_reading
    
    lgbm_hat = (gbm_regress.predict(data_4a2_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(lgbm_hat, data_4a2_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)
print('FinalScore:', sum(nrmlses) / K)

# Building_id mean by month
nrmlses = []
for tr_ix, ts_ix in kf.split(data_4a):
    # Building Means
    full_training_data = data_4a.iloc[tr_ix[:int(len(tr_ix)*.9)],:]
    building_id_means = full_training_data.groupby(['building_id', 'dt_m'], as_index=False).agg({'meter_reading': 'mean'})
    bid_dicts = {}
    for bid in list(set(building_id_means.building_id)):
        bid_data = building_id_means[building_id_means.building_id == bid][['dt_m', 'meter_reading']].to_dict('records')
        bid_dicts[bid] = bid_data
    building_id_stds = full_training_data.groupby(['building_id', 'dt_m'], as_index=False).agg({'meter_reading': 'std'})
    bid_dicts_s = {}
    for bids in list(set(building_id_stds.building_id)):
        bid_datas = building_id_stds[building_id_stds.building_id == bids][['dt_m', 'meter_reading']].to_dict('records')
        bid_dicts_s[bids] = bid_datas
    # Using Means / STDS
    bid_means_tr = pd.DataFrame(full_training_data.building_id.apply(lambda x: [x['meter_reading'] for x in bid_dicts[x]]).values.tolist(),
                      columns=['building_id_dt_m_{}_mean'.format(m) for m in [mm['dt_m'] for mm in bid_dicts[bid]]]).reset_index(drop=True)
    data_4a2_tr_x = full_training_data.reset_index(drop=True).merge(bid_means_tr, left_index=True, right_index=True)
    bid_std_tr = pd.DataFrame(full_training_data.building_id.apply(lambda x: [x['meter_reading'] for x in bid_dicts_s[x]]).values.tolist(),
                      columns=['building_id_dt_m_{}_std'.format(m) for m in [mm['dt_m'] for mm in bid_dicts_s[bids]]]).reset_index(drop=True)
    data_4a2_tr_x = data_4a2_tr_x.reset_index(drop=True).merge(bid_std_tr, left_index=True, right_index=True)
    data_4a2_tr_x = data_4a2_tr_x[[x for x in list(data_4a2_tr_x) if x not in ['meter_reading', 'building_id']]].fillna(0)
    data_4a2_tr_y = data_4a.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading
    
    data_4a2_val_x = data_4a.iloc[tr_ix[int(len(tr_ix)*.9):],:]
    bid_means_val = pd.DataFrame(data_4a2_val_x.building_id.apply(lambda x: [x['meter_reading'] for x in bid_dicts[x]]).values.tolist(),
                  columns=['building_id_dt_m_{}_mean'.format(m) for m in [mm['dt_m'] for mm in bid_dicts[bid]]]).reset_index(drop=True)
    data_4a2_val_x = data_4a2_val_x.reset_index(drop=True).merge(bid_means_val, left_index=True, right_index=True)
    bid_std_val = pd.DataFrame(data_4a2_val_x.building_id.apply(lambda x: [x['meter_reading'] for x in bid_dicts_s[x]]).values.tolist(),
                      columns=['building_id_dt_m_{}_std'.format(m) for m in [mm['dt_m'] for mm in bid_dicts_s[bids]]]).reset_index(drop=True)
    data_4a2_val_x = data_4a2_val_x.reset_index(drop=True).merge(bid_std_val, left_index=True, right_index=True)
    data_4a2_val_x = data_4a2_val_x[[x for x in list(data_4a2_val_x) if x not in ['meter_reading', 'building_id']]].fillna(0)
    data_4a2_val_y = data_4a.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading
    
    lgb_train = lgb.Dataset(data_4a2_tr_x, data_4a2_tr_y)
    lgb_eval = lgb.Dataset(data_4a2_val_x, data_4a2_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    
    data_4a2_ts_x = data_4a.iloc[ts_ix,:]
    bid_means_ts = pd.DataFrame(data_4a2_ts_x.building_id.apply(lambda x: [x['meter_reading'] for x in bid_dicts[x]]).values.tolist(),
                  columns=['building_id_dt_m_{}_mean'.format(m) for m in [mm['dt_m'] for mm in bid_dicts[bid]]]).reset_index(drop=True)
    data_4a2_ts_x = data_4a2_ts_x.reset_index(drop=True).merge(bid_means_ts, left_index=True, right_index=True)
    bid_std_val = pd.DataFrame(data_4a2_ts_x.building_id.apply(lambda x: [x['meter_reading'] for x in bid_dicts_s[x]]).values.tolist(),
                      columns=['building_id_dt_m_{}_std'.format(m) for m in [mm['dt_m'] for mm in bid_dicts_s[bids]]]).reset_index(drop=True)
    data_4a2_val_x = data_4a2_ts_x.reset_index(drop=True).merge(bid_std_val, left_index=True, right_index=True)
    data_4a2_ts_x = data_4a2_ts_x[[x for x in list(data_4a2_ts_x) if x not in ['meter_reading', 'building_id']]].fillna(0)
    data_4a2_ts_y = data_4a.iloc[ts_ix,:].meter_reading
    
    lgbm_hat = (gbm_regress.predict(data_4a2_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(lgbm_hat, data_4a2_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)
print('FinalScore:', sum(nrmlses) / K)
