==> nbs_arxiv/02-experiments-level5-site_id.ipynb <==
%load_ext autoreload
%autoreload 2

import pandas as pd
from tsm.data_selector import data_subset_by_dict
import matplotlib.pyplot as plt
import numpy as np

train_data = pd.read_pickle('kaggle/input/ashrae-energy-prediction/train.pkl')

train_data.primary_use.value_counts()

## Level 5A)
* 1 primary use
* 1 meter reading - meter 0
* Site ID

level_5a = {'primary_use': 1, 'meter': 0}
level_5a_cols = ['building_id', 'site_id', 'meter_reading', 'dt_m', 'dt_w', 'dt_d', 'dt_hour', 
                 'dt_day_week', 'dt_day_month', 'dt_week_month']

data_5a = data_subset_by_dict(train_data, level_5a, level_5a_cols)

data_5a.head()

len(set(data_5a.site_id))

len(set(data_5a.building_id))

import lightgbm as lgb

from tsm.eval_metrics import root_mean_squared_log_error

SEED = 7

lgb_reg_params = {'objective':'regression',  'boosting_type':'gbdt', 'metric':'rmse',
                  'n_jobs':-1, 'learning_rate':0.07, 'num_leaves': 2**8, 'max_depth':-1,
                  'tree_learner':'serial', 'colsample_bytree': 0.7, 'subsample_freq':1,
                  'subsample':0.5, 'max_bin':255, 'verbose':1, 'seed': SEED, 'early_stopping_rounds': 40} 

# Cross Validation
from sklearn.model_selection import KFold

K = 5
kf = KFold(n_splits=K)

feat_cols = ['building_id', 'site_id', 'dt_m', 'dt_w', 'dt_d', 'dt_hour', 
             'dt_day_week', 'dt_day_month', 'dt_week_month']

nrmlses = []
for tr_ix, ts_ix in kf.split(data_5a):
    data_5a1_tr_x = data_5a.iloc[tr_ix[:int(len(tr_ix)*.9)],:][feat_cols]
    data_5a1_tr_y = data_5a.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading
    data_5a1_val_x = data_5a.iloc[tr_ix[int(len(tr_ix)*.9):],:][feat_cols]
    data_5a1_val_y = data_5a.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading
    lgb_train = lgb.Dataset(data_5a1_tr_x, data_5a1_tr_y)
    lgb_eval = lgb.Dataset(data_5a1_val_x, data_5a1_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    data_5a1_ts_x = data_5a.iloc[ts_ix,:][feat_cols]
    data_5a1_ts_y = data_5a.iloc[ts_ix,:].meter_reading
    lgbm_hat = (gbm_regress.predict(data_5a1_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(lgbm_hat, data_5a1_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)
print('FinalScore:', sum(nrmlses) / K)

feat_cols = ['building_id', 'site_id_mean', 'dt_m', 'dt_w', 'dt_d', 'dt_hour', 
             'dt_day_week', 'dt_day_month', 'dt_week_month']

# Simple building_id mean
nrmlses = []
for tr_ix, ts_ix in kf.split(data_5a):
    full_training_data = data_5a.iloc[tr_ix[:int(len(tr_ix)*.9)],:]
    site_id_means = full_training_data.groupby('site_id')['meter_reading'].mean().to_dict()
    full_training_data['site_id_mean'] = full_training_data.site_id.apply(lambda x: site_id_means[x])
    data_5a2_tr_x = full_training_data[feat_cols]
    data_5a2_tr_y = data_5a.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading
    data_5a2_val_x = data_5a.iloc[tr_ix[int(len(tr_ix)*.9):],:]
    data_5a2_val_x['site_id_mean'] = data_5a2_val_x.site_id.apply(lambda x: site_id_means[x])
    data_5a2_val_x = data_5a2_val_x[feat_cols]
    data_5a2_val_y = data_5a.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading
    lgb_train = lgb.Dataset(data_5a2_tr_x, data_5a2_tr_y)
    lgb_eval = lgb.Dataset(data_5a2_val_x, data_5a2_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    data_5a2_ts_x = data_5a.iloc[ts_ix,:]
    data_5a2_ts_x['site_id_mean'] = data_5a2_ts_x.site_id.apply(lambda x: site_id_means[x])
    data_5a2_ts_x = data_5a2_ts_x[feat_cols]
    data_5a2_ts_y = data_5a.iloc[ts_ix,:].meter_reading
    lgbm_hat = (gbm_regress.predict(data_5a2_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(lgbm_hat, data_5a2_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)
print('FinalScore:', sum(nrmlses) / K)

# Site_id mean by month
nrmlses = []
for tr_ix, ts_ix in kf.split(data_5a):
    # Building Means
    full_training_data = data_5a.iloc[tr_ix[:int(len(tr_ix)*.9)],:]
    site_id_means = full_training_data.groupby(['site_id', 'dt_m'], as_index=False)['meter_reading'].mean()
    sid_dicts = {}
    for sid in list(set(site_id_means.site_id)):
        sid_data = site_id_means[site_id_means.site_id == sid][['dt_m', 'meter_reading']].to_dict('records')
        sid_dicts[sid] = sid_data
    # Using Means
    sid_means_tr = pd.DataFrame(full_training_data.site_id.apply(lambda x: [x['meter_reading'] for x in sid_dicts[x]]).values.tolist(),
                      columns=['site_id_dt_m_{}_mean'.format(m) for m in [mm['dt_m'] for mm in sid_dicts[sid]]]).reset_index(drop=True)
    data_4a2_tr_x = full_training_data.reset_index(drop=True).merge(sid_means_tr, left_index=True, right_index=True)
    data_4a2_tr_x = data_4a2_tr_x[[x for x in list(data_4a2_tr_x) if x not in ['meter_reading', 'site_id']]].fillna(0)
    data_4a2_tr_y = data_5a.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading
    
    data_4a2_val_x = data_5a.iloc[tr_ix[int(len(tr_ix)*.9):],:]
    bid_means_val = pd.DataFrame(data_4a2_val_x.site_id.apply(lambda x: [x['meter_reading'] for x in sid_dicts[x]]).values.tolist(),
                  columns=['site_id_dt_m_{}_mean'.format(m) for m in [mm['dt_m'] for mm in sid_dicts[sid]]]).reset_index(drop=True)
    data_4a2_val_x = data_4a2_val_x.reset_index(drop=True).merge(bid_means_val, left_index=True, right_index=True)
    data_4a2_val_x = data_4a2_val_x[[x for x in list(data_4a2_val_x) if x not in ['meter_reading', 'site_id']]].fillna(0)
    data_4a2_val_y = data_5a.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading
    
    lgb_train = lgb.Dataset(data_4a2_tr_x, data_4a2_tr_y)
    lgb_eval = lgb.Dataset(data_4a2_val_x, data_4a2_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    
    data_4a2_ts_x = data_5a.iloc[ts_ix,:]
    bid_means_ts = pd.DataFrame(data_4a2_ts_x.site_id.apply(lambda x: [x['meter_reading'] for x in sid_dicts[x]]).values.tolist(),
                  columns=['site_id_dt_m_{}_mean'.format(m) for m in [mm['dt_m'] for mm in sid_dicts[sid]]]).reset_index(drop=True)
    data_4a2_ts_x = data_4a2_ts_x.reset_index(drop=True).merge(bid_means_ts, left_index=True, right_index=True)
    data_4a2_ts_x = data_4a2_ts_x[[x for x in list(data_4a2_ts_x) if x not in ['meter_reading', 'site_id']]].fillna(0)
    data_4a2_ts_y = data_5a.iloc[ts_ix,:].meter_reading
    
    lgbm_hat = (gbm_regress.predict(data_4a2_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(lgbm_hat, data_4a2_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)
print('FinalScore:', sum(nrmlses) / K)
