==> nbs_arxiv/02-experiments-level6-opt-lags.ipynb <==
%load_ext autoreload
%autoreload 2

import pandas as pd
from tsm.data_selector import data_subset_by_dict
import matplotlib.pyplot as plt
import numpy as np

train_data = pd.read_pickle('kaggle/input/ashrae-energy-prediction/train.pkl')

## Level 6A)
* 1 use: 4\
* 1 meter reading - meter 0
* Weather features

level_6a = {'primary_use': 4, 'meter': 0}
level_6a_cols = ['building_id', 'meter_reading', 'air_temperature', 'cloud_coverage', 'dew_temperature', 'precip_depth_1_hr', 'sea_level_pressure', 'wind_direction','wind_speed']

data_6a = data_subset_by_dict(train_data, level_6a, level_6a_cols)

# Many buildings?
len(set(data_6a.building_id))

data_6a.meter_reading.ewm(com=0.8).mean().tail()

import lightgbm as lgb

from tsm.eval_metrics import root_mean_squared_log_error

SEED = 7

lgb_reg_params = {'objective':'regression',  'boosting_type':'gbdt', 'metric':'rmse',
                  'n_jobs':-1, 'learning_rate':0.07, 'num_leaves': 2**8, 'max_depth':-1,
                  'tree_learner':'serial', 'colsample_bytree': 0.7, 'subsample_freq':1,
                  'subsample':0.5, 'max_bin':255, 'verbose':1, 'seed': SEED, 'early_stopping_rounds': 40} 

# Cross Validation
from sklearn.model_selection import KFold

K = 5
kf = KFold(n_splits=K)

feat_cols = [x for x in level_6a_cols if x != 'meter_reading']

nrmlses = []
for tr_ix, ts_ix in kf.split(data_6a):
    data_6a1_tr_x = data_6a.iloc[tr_ix[:int(len(tr_ix)*.9)],:][feat_cols]
    data_6a1_tr_y = data_6a.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading
    data_6a1_val_x = data_6a.iloc[tr_ix[int(len(tr_ix)*.9):],:][feat_cols]
    data_6a1_val_y = data_6a.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading
    lgb_train = lgb.Dataset(data_6a1_tr_x, data_6a1_tr_y)
    lgb_eval = lgb.Dataset(data_6a1_val_x, data_6a1_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    data_6a1_ts_x = data_6a.iloc[ts_ix,:][feat_cols]
    data_6a1_ts_y = data_6a.iloc[ts_ix,:].meter_reading
    lgbm_hat = (gbm_regress.predict(data_6a1_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(lgbm_hat, data_6a1_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)
print('FinalScore:', sum(nrmlses) / K)

from tsm.data_selector import get_series_past_k_lags

nrmlses = []
for tr_ix, ts_ix in kf.split(data_6a):
    k_air = get_series_past_k_lags(data_6a['air_temperature'], k=[3 * (2**x) for x in range(6)]).fillna(99).reset_index(drop=True)
    k_dew = get_series_past_k_lags(data_6a['dew_temperature'], k=[3 * (2**x) for x in range(6)]).fillna(99).reset_index(drop=True)
    data_6a2 = pd.concat([data_6a.reset_index(drop=True), k_air, k_dew], axis=1, sort=False)
    data_6a2_tr_x = data_6a2.iloc[tr_ix[:int(len(tr_ix)*.9)],:][[x for x in list(data_6a2) if x != 'meter_reading']]
    data_6a2_tr_y = data_6a2.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading
    data_6a2_val_x = data_6a2.iloc[tr_ix[int(len(tr_ix)*.9):],:][[x for x in list(data_6a2) if x != 'meter_reading']]
    data_6a2_val_y = data_6a2.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading
    lgb_train = lgb.Dataset(data_6a2_tr_x, data_6a2_tr_y)
    lgb_eval = lgb.Dataset(data_6a2_val_x, data_6a2_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    data_6a2_ts_x = data_6a2.iloc[ts_ix,:][[x for x in list(data_6a2) if x != 'meter_reading']]
    data_6a2_ts_y = data_6a2.iloc[ts_ix,:].meter_reading
    lgbm_hat = (gbm_regress.predict(data_6a2_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(lgbm_hat, data_6a2_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)
print('FinalScore:', sum(nrmlses) / K)

nrmlses = []
for tr_ix, ts_ix in kf.split(data_6a):
    k_air_e_001 = data_6a.air_temperature.ewm(alpha=0.01).mean().reset_index(drop=True)
    k_air_e_001.name = 'k_air_e_001'
    k_dew_e_001 = data_6a.dew_temperature.ewm(alpha=0.01).mean().reset_index(drop=True)
    k_dew_e_001.name = 'k_dew_e_001'
    k_air_e_01 = data_6a.air_temperature.ewm(alpha=0.1).mean().reset_index(drop=True)
    k_air_e_01.name = 'k_air_e_01'
    k_dew_e_01 = data_6a.dew_temperature.ewm(alpha=0.1).mean().reset_index(drop=True)
    k_dew_e_01.name = 'k_dew_e_01'
    k_air_e_025 = data_6a.air_temperature.ewm(alpha=0.25).mean().reset_index(drop=True)
    k_air_e_025.name = 'k_air_e_025'
    k_dew_e_025 = data_6a.dew_temperature.ewm(alpha=0.25).mean().reset_index(drop=True)
    k_dew_e_025.name = 'k_dew_e_025'
    data_6a3 = pd.concat([data_6a.reset_index(drop=True), k_air_e_001,k_air_e_01, k_air_e_025, k_dew_e_001, k_dew_e_01, k_dew_e_025], axis=1, sort=False)
    data_6a3_tr_x = data_6a3.iloc[tr_ix[:int(len(tr_ix)*.9)],:][[x for x in list(data_6a3) if x != 'meter_reading']]
    data_6a3_tr_y = data_6a3.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading
    data_6a3_val_x = data_6a3.iloc[tr_ix[int(len(tr_ix)*.9):],:][[x for x in list(data_6a3) if x != 'meter_reading']]
    data_6a3_val_y = data_6a3.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading
    lgb_train = lgb.Dataset(data_6a3_tr_x, data_6a3_tr_y)
    lgb_eval = lgb.Dataset(data_6a3_val_x, data_6a3_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    data_6a3_ts_x = data_6a3.iloc[ts_ix,:][[x for x in list(data_6a3) if x != 'meter_reading']]
    data_6a3_ts_y = data_6a3.iloc[ts_ix,:].meter_reading
    lgbm_hat = (gbm_regress.predict(data_6a3_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(lgbm_hat, data_6a3_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)
print('FinalScore:', sum(nrmlses) / K)

## Level 6B)
* 1 use: 0
* 1 meter reading - meter 1
* Weather features

level_6a = {'primary_use': 0, 'meter': 1}
level_6a_cols = ['building_id', 'site_id', 'meter_reading', 'air_temperature', 'cloud_coverage', 'dew_temperature', 'precip_depth_1_hr', 'sea_level_pressure', 'wind_direction','wind_speed']

data_6a = data_subset_by_dict(train_data, level_6a, level_6a_cols)

# Many buildings?
len(set(data_6a.building_id))

feat_cols = [x for x in level_6a_cols if x != 'meter_reading']

nrmlses = []
for tr_ix, ts_ix in kf.split(data_6a):
    data_6a1_tr_x = data_6a.iloc[tr_ix[:int(len(tr_ix)*.9)],:][feat_cols]
    data_6a1_tr_y = data_6a.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading
    data_6a1_val_x = data_6a.iloc[tr_ix[int(len(tr_ix)*.9):],:][feat_cols]
    data_6a1_val_y = data_6a.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading
    lgb_train = lgb.Dataset(data_6a1_tr_x, data_6a1_tr_y)
    lgb_eval = lgb.Dataset(data_6a1_val_x, data_6a1_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    data_6a1_ts_x = data_6a.iloc[ts_ix,:][feat_cols]
    data_6a1_ts_y = data_6a.iloc[ts_ix,:].meter_reading
    lgbm_hat = (gbm_regress.predict(data_6a1_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(lgbm_hat, data_6a1_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)
print('FinalScore:', sum(nrmlses) / K)

nrmlses = []
for tr_ix, ts_ix in kf.split(data_6a):
    k_air = get_series_past_k_lags(data_6a['air_temperature'], k=[1, 2] + [3 * (2**x) for x in range(6)]).fillna(99).reset_index(drop=True)
    k_dew = get_series_past_k_lags(data_6a['dew_temperature'], k=[1, 2] + [3 * (2**x) for x in range(6)]).fillna(99).reset_index(drop=True)
    data_6a2 = pd.concat([data_6a.reset_index(drop=True), k_air, k_dew], axis=1, sort=False)
    data_6a2_tr_x = data_6a2.iloc[tr_ix[:int(len(tr_ix)*.9)],:][[x for x in list(data_6a2) if x != 'meter_reading']]
    data_6a2_tr_y = data_6a2.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading
    data_6a2_val_x = data_6a2.iloc[tr_ix[int(len(tr_ix)*.9):],:][[x for x in list(data_6a2) if x != 'meter_reading']]
    data_6a2_val_y = data_6a2.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading
    lgb_train = lgb.Dataset(data_6a2_tr_x, data_6a2_tr_y)
    lgb_eval = lgb.Dataset(data_6a2_val_x, data_6a2_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    data_6a2_ts_x = data_6a2.iloc[ts_ix,:][[x for x in list(data_6a2) if x != 'meter_reading']]
    data_6a2_ts_y = data_6a2.iloc[ts_ix,:].meter_reading
    lgbm_hat = (gbm_regress.predict(data_6a2_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(lgbm_hat, data_6a2_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)
print('FinalScore:', sum(nrmlses) / K)

nrmlses = []
for tr_ix, ts_ix in kf.split(data_6a):
    k_air = get_series_past_k_lags(data_6a['air_temperature'], k=[1, 2, 3, 6]).fillna(99).reset_index(drop=True)
    k_dew = get_series_past_k_lags(data_6a['dew_temperature'], k=[1, 2, 3, 6]).fillna(99).reset_index(drop=True)
    k_air_e_01 = data_6a.air_temperature.ewm(alpha=0.1).mean().reset_index(drop=True)
    k_air_e_01.name = 'k_air_e_01'
    k_dew_e_01 = data_6a.dew_temperature.ewm(alpha=0.1).mean().reset_index(drop=True)
    k_dew_e_01.name = 'k_dew_e_01'
    k_air_e_025 = data_6a.air_temperature.ewm(alpha=0.25).mean().reset_index(drop=True)
    k_air_e_025.name = 'k_air_e_025'
    k_dew_e_025 = data_6a.dew_temperature.ewm(alpha=0.25).mean().reset_index(drop=True)
    k_dew_e_025.name = 'k_dew_e_025'
    data_6a3 = pd.concat([data_6a.reset_index(drop=True), k_air, k_air_e_001,k_air_e_01, k_air_e_025, k_dew, k_dew_e_001, k_dew_e_01, k_dew_e_025], axis=1, sort=False)
    data_6a3_tr_x = data_6a3.iloc[tr_ix[:int(len(tr_ix)*.9)],:][[x for x in list(data_6a3) if x != 'meter_reading']]
    data_6a3_tr_y = data_6a3.iloc[tr_ix[:int(len(tr_ix)*.9)],:].meter_reading
    data_6a3_val_x = data_6a3.iloc[tr_ix[int(len(tr_ix)*.9):],:][[x for x in list(data_6a3) if x != 'meter_reading']]
    data_6a3_val_y = data_6a3.iloc[tr_ix[int(len(tr_ix)*.9):],:].meter_reading
    lgb_train = lgb.Dataset(data_6a3_tr_x, data_6a3_tr_y)
    lgb_eval = lgb.Dataset(data_6a3_val_x, data_6a3_val_y)
    gbm_regress = lgb.train(lgb_reg_params, lgb_train, num_boost_round=2000, valid_sets=(lgb_train, lgb_eval),
                            early_stopping_rounds=20, verbose_eval = 20)
    data_6a3_ts_x = data_6a3.iloc[ts_ix,:][[x for x in list(data_6a3) if x != 'meter_reading']]
    data_6a3_ts_y = data_6a3.iloc[ts_ix,:].meter_reading
    lgbm_hat = (gbm_regress.predict(data_6a3_ts_x, num_iteration=gbm_regress.best_iteration))
    nrmlse = root_mean_squared_log_error(lgbm_hat, data_6a3_ts_y)
    print('Iteration done, nrmlse:', nrmlse)
    nrmlses.append(nrmlse)
print('FinalScore:', sum(nrmlses) / K)
