==> nbs_arxiv/03-preparation+train_val.ipynb <==
%load_ext autoreload
%autoreload 2

# Preparing Data for Training

import pandas as pd
from tsm.data_utils import time_processing, encode_categories, ordinal2wave
from tsm.data_selector import get_series_past_k_lags

train_data = pd.read_pickle('kaggle/train_data.pkl')

train_data.head()

## Step 1) Add ordinal2wave for temporal features

from tsm.data_utils import ordinal2wave

train_data = time_processing(train_data, 'timestamp')

train_data, econder = encode_categories(train_data, cat_cols='primary_use')

from sklearn.externals import joblib

joblib.dump(econder, 'models/LabelEconder_train_data.pkl')

train_data = ordinal2wave('dt_m', train_data)
train_data = ordinal2wave('dt_w', train_data)
train_data = ordinal2wave('dt_d', train_data)
train_data = ordinal2wave('dt_hour', train_data)
train_data = ordinal2wave('dt_day_week', train_data)
train_data = ordinal2wave('dt_day_month', train_data)
train_data = ordinal2wave('dt_week_month', train_data)

## Step 2) Add weather lags (mainly air and dew temperature) Alpha 00.1, 0.1, 0.25

train_data_pp = pd.read_pickle('kaggle/train_processed.pkl')

k_air_e_001 = train_data.air_temperature.ewm(alpha=0.01).mean().reset_index(drop=True)
k_air_e_001.name = 'k_air_e_001'

train_data_pp = pd.concat([train_data_pp.reset_index(drop=True), k_air_e_001], axis=1, sort=False)

del k_air_e_001

k_air_e_01 = train_data.air_temperature.ewm(alpha=0.1).mean().reset_index(drop=True)
k_air_e_01.name = 'k_air_e_01'

train_data_pp = pd.concat([train_data_pp.reset_index(drop=True), k_air_e_01], axis=1, sort=False)

del k_air_e_01

k_air_e_025 = train_data.air_temperature.ewm(alpha=0.25).mean().reset_index(drop=True)
k_air_e_025.name = 'k_air_e_025'

train_data_pp = pd.concat([train_data_pp.reset_index(drop=True), k_air_e_025], axis=1, sort=False)

del k_air_e_025

train_data_pp.to_pickle('kaggle/train_processed.pkl')

train_data_pp.head()

from tsm.data_selector import data_subset_by_dict

## Step3) Sample validation set

pd.set_option('display.max_columns', 100)

weeks_no_and_of_m = train_data_pp[['dt_w', 'dt_m', 'dt_week_month']]

only_even_months = weeks_no_and_of_m[weeks_no_and_of_m.dt_m % 2 == 0]

validation_weeks = []
for m in list(set(only_even_months.dt_m)):
    m_data = only_even_months[only_even_months.dt_m == m]
    m_data_sample = m_data.sample(n=1, axis=0)
    print(m_data_sample)
    validation_weeks.append(m_data_sample.dt_w.values[0])

validation_weeks

training_weeks  = [x for x in range(1, 54) if x not in validation_weeks]

prepo_train = train_data_pp[train_data_pp.dt_w.isin(training_weeks)]
prepo_val = train_data_pp[train_data_pp.dt_w.isin(validation_weeks)]

prepo_train.to_pickle('kaggle/preproc_train.pkl')

prepo_val.to_pickle('kaggle/preproc_val.pkl')

del train_data_pp

## Step4) Divide by meter type

from tsm.data_selector import data_subset_by_dict

### Meter 0

prepo_train_met0 = data_subset_by_dict(prepo_train, {'meter': 0})
prepo_val_met0 = data_subset_by_dict(prepo_val, {'meter': 0})

prepo_train_met0.to_pickle('kaggle/prepo_train_met0.pkl')
prepo_val_met0.to_pickle('kaggle/prepo_val_met0.pkl')

del prepo_train_met0
del prepo_val_met0

### Meter 1

prepo_train_met1 = data_subset_by_dict(prepo_train, {'meter': 1})
prepo_val_met1 = data_subset_by_dict(prepo_val, {'meter': 1})

prepo_train_met1.to_pickle('kaggle/prepo_train_met1.pkl')
prepo_val_met1.to_pickle('kaggle/prepo_val_met1.pkl')

del prepo_train_met1
del prepo_val_met1

### Meter 2

prepo_train_met2 = data_subset_by_dict(prepo_train, {'meter': 2})
prepo_val_met2 = data_subset_by_dict(prepo_val, {'meter': 2})

prepo_train_met2.to_pickle('kaggle/prepo_train_met2.pkl')
prepo_val_met2.to_pickle('kaggle/prepo_val_met2.pkl')

del prepo_train_met2
del prepo_val_met2

### Meter 3

prepo_train_met3 = data_subset_by_dict(prepo_train, {'meter': 3})
prepo_val_met3 = data_subset_by_dict(prepo_val, {'meter': 3})

prepo_train_met3.to_pickle('kaggle/prepo_train_met3.pkl')
prepo_val_met3.to_pickle('kaggle/prepo_val_met3.pkl')
