==> nbs_arxiv/04-training-baselines.ipynb <==
# Training Models Baselines

%load_ext autoreload
%autoreload 2

import pandas as pd

## Meter 0

met0_tr = pd.read_pickle('kaggle/prepo_train_met0.pkl')
met0_vl = pd.read_pickle('kaggle/prepo_val_met0.pkl')

len(met0_tr), len(met0_vl)

### Compute mean meter reading by building by month on training data

from tsm.data_utils import compute_and_add_mean_by_a_and_b, only_add_mean_by_a_and_b

met0_tr, tr_means_by_bid_and_m = compute_and_add_mean_by_a_and_b(met0_tr, mean_of='meter_reading', a='building_id', b='dt_m')

met0_vl = only_add_mean_by_a_and_b(met0_vl, mean_of='meter_reading', a='building_id', b='dt_m', means_by_a_and_b=tr_means_by_bid_and_m)

#### Re-compress memory usage

from tsm.data_utils import compress_memory_usage

met0_tr = met0_tr.drop('timestamp', axis=1)

met0_tr, _ = compress_memory_usage(met0_tr)

met0_vl = met0_vl.drop('timestamp', axis=1)

met0_vl, _ = compress_memory_usage(met0_vl)

#### Save as pickle to save time later

met0_tr.to_pickle('kaggle/prepo_train_met0_bIDbyM_means.pkl')
met0_vl.to_pickle('kaggle/prepo_val_met0__bIDbyM_means.pkl')

import pickle

with open('kaggle/prepo_train_met0_tr_means_by_bid_and_m.pkl', 'wb') as f_out:
    pickle.dump(tr_means_by_bid_and_m, f_out)

#### Reloading: restart from here

met0_tr = pd.read_pickle('kaggle/prepo_train_met0_bIDbyM_means.pkl')
met0_vl = pd.read_pickle('kaggle/prepo_val_met0__bIDbyM_means.pkl')

with open('kaggle/prepo_train_met0_tr_means_by_bid_and_m.pkl', 'rb') as f_in:
    mt0_tr_means_by_bid_and_m = pickle.load(f_in)

## Baseline 1) LightGBM

import lightgbm as lgb

from tsm.eval_metrics import root_mean_squared_log_error

SEED = 0

lgb_reg_params = {'objective':'regression',  'boosting_type':'gbdt', 'metric':'rmse',
                  'n_jobs':-1, 'learning_rate':0.07, 'num_leaves': 2**8, 'max_depth':-1,
                  'tree_learner':'serial', 'colsample_bytree': 0.7, 'subsample_freq':1,
                  'subsample':0.5, 'max_bin':255, 'verbose':1, 'seed': SEED, 'early_stopping_rounds': 40}

xs = [x for x in list(met0_tr) if x not in ['timestamp', 'meter_reading']]
y = 'meter_reading'

met0_tr_x = met0_tr[xs]
met0_tr_y = met0_tr[y]
met0_vl_x = met0_vl[xs]
met0_vl_y = met0_vl[y]

%whos

del met0_tr, met0_vl

mt0_lgb_train = lgb.Dataset(met0_tr_x, met0_tr_y)
mt0_lgb_eval = lgb.Dataset(met0_vl_x, met0_vl_y)
mt0_gbm_regress = lgb.train(lgb_reg_params, mt0_lgb_train, num_boost_round=2000, valid_sets=(mt0_lgb_train, mt0_lgb_eval),verbose_eval = 20)

lgbm_tr_hat = (gbm_regress.predict(met0_tr_x, num_iteration=gbm_regress.best_iteration))
print('Training NRMLSE:', root_mean_squared_log_error(lgbm_tr_hat, met0_tr_y))

lgbm_vl_hat = (gbm_regress.predict(met0_vl_x, num_iteration=gbm_regress.best_iteration))
print('Validation NRMLSE:', root_mean_squared_log_error(lgbm_vl_hat, met0_vl_y))

from sklearn.externals import joblib

joblib.dump(mt0_gbm_regress, 'models/mt0_gbm_regress.pkl')

## Meter 1

met1_tr = pd.read_pickle('kaggle/prepo_train_met1.pkl')
met1_vl = pd.read_pickle('kaggle/prepo_val_met1.pkl')

met1_tr, mt_1tr_means_by_bid_and_m = compute_and_add_mean_by_a_and_b(met1_tr, mean_of='meter_reading', a='building_id', b='dt_m')

met1_vl = only_add_mean_by_a_and_b(met1_vl, mean_of='meter_reading', a='building_id', b='dt_m', means_by_a_and_b=mt_1tr_means_by_bid_and_m)

met1_tr = met1_tr.drop('timestamp', axis=1)

met1_tr, _ = compress_memory_usage(met1_tr)

met1_vl = met1_vl.drop('timestamp', axis=1)

met1_vl, _ = compress_memory_usage(met1_vl)

met1_tr.to_pickle('kaggle/prepo_train_met1_bIDbyM_means.pkl')
met1_vl.to_pickle('kaggle/prepo_val_met1_bIDbyM_means.pkl')

with open('kaggle/prepo_train_met1_tr_means_by_bid_and_m.pkl', 'wb') as f_out:
    pickle.dump(mt_1tr_means_by_bid_and_m, f_out)

met1_tr = pd.read_pickle('kaggle/prepo_train_met1_bIDbyM_means.pkl')
met1_vl = pd.read_pickle ('kaggle/prepo_val_met1_bIDbyM_means.pkl')

xs = [x for x in list(met1_tr) if x not in ['timestamp', 'meter_reading']]
y = 'meter_reading'

met1_tr_x = met1_tr[xs]
met1_tr_y = met1_tr[y]
met1_vl_x = met1_vl[xs]
met1_vl_y = met1_vl[y]

mt1_lgb_train = lgb.Dataset(met1_tr_x, met1_tr_y)
mt1_lgb_eval = lgb.Dataset(met1_vl_x, met1_vl_y)
mt1_gbm_regress = lgb.train(lgb_reg_params, mt1_lgb_train, num_boost_round=2000, valid_sets=(mt1_lgb_train, mt1_lgb_eval),verbose_eval = 20)

lgbm_tr_hat = (gbm_regress.predict(met1_tr_x, num_iteration=gbm_regress.best_iteration))
print('Training NRMLSE:', root_mean_squared_log_error(lgbm_tr_hat, met1_tr_y))

lgbm_vl_hat = (gbm_regress.predict(met1_vl_x, num_iteration=gbm_regress.best_iteration))
print('Validation NRMLSE:', root_mean_squared_log_error(lgbm_vl_hat, met1_vl_y))

joblib.dump(mt1_gbm_regress, 'models/mt1_gbm_regress.pkl')

%whos

del mt0_tr_means_by_bid_and_m

del lgb_train, lgb_eval, lgbm_tr_hat, lgbm_vl_hat, met1_tr, met1_tr_x, met1_tr_y, met1_vl, met1_vl_x, met1_vl_y, mt_1tr_means_by_bid_and_m

del met1_tr, met1_vl

## Meter 2

met2_tr = pd.read_pickle('kaggle/prepo_train_met2.pkl')
met2_vl = pd.read_pickle('kaggle/prepo_val_met2.pkl')

met2_tr, mt_2tr_means_by_bid_and_m = compute_and_add_mean_by_a_and_b(met2_tr, mean_of='meter_reading', a='building_id', b='dt_m')

met2_vl = only_add_mean_by_a_and_b(met2_vl, mean_of='meter_reading', a='building_id', b='dt_m', means_by_a_and_b=mt_2tr_means_by_bid_and_m)

met2_tr = met2_tr.drop('timestamp', axis=1)

met2_tr, _ = compress_memory_usage(met2_tr)

met2_vl = met2_vl.drop('timestamp', axis=1)

met2_vl, _ = compress_memory_usage(met2_vl)

met2_tr = pd.read_pickle('kaggle/prepo_train_met2_bIDbyM_means.pkl')
met2_vl = pd.read_pickle('kaggle/prepo_val_met2_bIDbyM_means.pkl')

with open('kaggle/prepo_train_met2_tr_means_by_bid_and_m.pkl', 'wb') as f_out:
    pickle.dump(mt_2tr_means_by_bid_and_m, f_out)

xs = [x for x in list(met2_tr) if x not in ['timestamp', 'meter_reading']]
y = 'meter_reading'

met2_tr_x = met2_tr[xs]
met2_tr_y = met2_tr[y]
met2_vl_x = met2_vl[xs]
met2_vl_y = met2_vl[y]

mt2_lgb_train = lgb.Dataset(met2_tr_x, met2_tr_y)
mt2_lgb_eval = lgb.Dataset(met2_vl_x, met2_vl_y)
mt2_gbm_regress = lgb.train(lgb_reg_params, mt2_lgb_train, num_boost_round=2000, valid_sets=(mt2_lgb_train, mt2_lgb_eval),verbose_eval = 20)

lgbm_tr_hat = (gbm_regress.predict(met2_tr_x, num_iteration=gbm_regress.best_iteration))
print('Training NRMLSE:', root_mean_squared_log_error(lgbm_tr_hat, met2_tr_y))

lgbm_vl_hat = (gbm_regress.predict(met2_vl_x, num_iteration=gbm_regress.best_iteration))
print('Validation NRMLSE:', root_mean_squared_log_error(lgbm_vl_hat, met2_vl_y))

joblib.dump(mt2_gbm_regress, 'models/mt2_gbm_regress.pkl')

%whos

del xs, y, mt_2tr_means_by_bid_and_m, met1_v2, met2_tr, met2_tr_x, met2_tr_y, met2_vl, met2_vl_x, met2_vl_y, lgb_train, lgbm_tr_hat, lgb_eval, lgbm_vl_hat

## Meter 3

met3_tr = pd.read_pickle('kaggle/prepo_train_met3.pkl')
met3_vl = pd.read_pickle('kaggle/prepo_val_met3.pkl')

met3_tr, mt_3tr_means_by_bid_and_m = compute_and_add_mean_by_a_and_b(met3_tr, mean_of='meter_reading', a='building_id', b='dt_m')

met3_vl = only_add_mean_by_a_and_b(met3_vl, mean_of='meter_reading', a='building_id', b='dt_m', means_by_a_and_b=mt_3tr_means_by_bid_and_m)

met3_tr = met3_tr.drop('timestamp', axis=1)

met3_tr, _ = compress_memory_usage(met3_tr)

met3_vl = met3_vl.drop('timestamp', axis=1)

met3_vl, _ = compress_memory_usage(met3_vl)

met3_tr = pd.read_pickle('kaggle/prepo_train_met3_bIDbyM_means.pkl')
met3_vl = pd.read_pickle('kaggle/prepo_val_met3_bIDbyM_means.pkl')

with open('kaggle/prepo_train_met3_tr_means_by_bid_and_m.pkl', 'wb') as f_out:
    pickle.dump(mt_3tr_means_by_bid_and_m, f_out)

xs = [x for x in list(met3_tr) if x not in ['timestamp', 'meter_reading']]
y = 'meter_reading'

met3_tr_x = met3_tr[xs]
met3_tr_y = met3_tr[y]
met3_vl_x = met3_vl[xs]
met3_vl_y = met3_vl[y]

mt3_lgb_train = lgb.Dataset(met3_tr_x, met3_tr_y)
mt3_lgb_eval = lgb.Dataset(met3_vl_x, met3_vl_y)
mt3_gbm_regress = lgb.train(lgb_reg_params, mt3_lgb_train, num_boost_round=2000, valid_sets=(mt3_lgb_train, mt3_lgb_eval),verbose_eval = 20)

lgbm_tr_hat = (gbm_regress.predict(met3_tr_x, num_iteration=gbm_regress.best_iteration))
print('Training NRMLSE:', root_mean_squared_log_error(lgbm_tr_hat, met3_tr_y))

lgbm_vl_hat = (gbm_regress.predict(met3_vl_x, num_iteration=gbm_regress.best_iteration))
print('Validation NRMLSE:', root_mean_squared_log_error(lgbm_vl_hat, met3_vl_y))

joblib.dump(mt3_gbm_regress, 'models/mt3_gbm_regress.pkl')

## Computing overall baselines

len_T0 = 10661397
len_V0 = 1399513
acc_T0 = 0.6020488382795164
acc_V0 = 0.6820828571711314
len_T1 = 3702435
len_V1 = 480005
acc_T1 = 1.5414553172411622
acc_V1 = 1.7537242860741826
len_T2 = 2387956
len_V2 = 320757
acc_T2 = 1.830378535525465
acc_V2 = 1.9614942522066001
len_T3 = 1118728
len_V3 = 145309
acc_T3 = 1.9423154769341195
acc_V3 = 2.1001962268893863

len_T = len_T0 + len_T1 + len_T2 + len_T3
len_V = len_V0 + len_V1 + len_V2 + len_V3

acc_T = acc_T0 * (len_T0 / len_T) + acc_T1 * (len_T1 / len_T) + acc_T2 * (len_T2 / len_T) + acc_T3 * (len_T3 / len_T)

acc_V = acc_V0 * (len_V0 / len_V) + acc_V1 * (len_V1 / len_V) + acc_V2 * (len_V2 / len_V) + acc_V3 * (len_V3 / len_V)

acc_T, acc_V

L = len_T + len_V

acc_T * len_T / L + acc_V * len_V / L
