==> nbs_arxiv/04-training-DNN.ipynb <==
# Deep Neural Networks

## Load Data

%load_ext autoreload
%autoreload 2

import pandas as pd

met0_tr = pd.read_pickle('kaggle/prepo_train_met0_bIDbyM_means.pkl')
met0_vl = pd.read_pickle('kaggle/prepo_val_met0__bIDbyM_means.pkl')

xs = [x for x in list(met0_tr) if x not in ['timestamp', 'meter_reading']]
y = 'meter_reading'

met0_tr_x = met0_tr[xs]
met0_tr_y = met0_tr[y]
met0_vl_x = met0_vl[xs]
met0_vl_y = met0_vl[y]

del met0_tr, met0_vl

## Build Network

import tensorflow as tf
from keras import layers

import keras

tf.reset_default_graph()

def build_model(no_of_feat):
    model = keras.Sequential([
        layers.Dense(64, activation='relu', input_shape=[no_of_feat]),
        layers.Dense(64, activation='relu'),
        layers.Dense(1, activation='relu')
    ])
    model.summary()
    model.compile(loss=root_mean_squared_logarithmic_error, optimizer='rmsprop', metrics=['mse'])
    return model

import keras.backend as K

def root_mean_squared_logarithmic_error(y_true, y_pred):
        return K.sqrt(K.mean(K.square(tf.math.log1p(y_pred) - tf.math.log1p(y_true))))

model = build_model(len(list(met0_tr_x)))

early_stopper = keras.callbacks.EarlyStopping(patience=20, verbose=1)
check_pointer = keras.callbacks.ModelCheckpoint('models/DNN_MT0_261019T1500.json', verbose=1, save_best_only=True)

model.fit(met0_tr_x, met0_tr_y, epochs=10, verbose=1, batch_size=256,
          validation_data=[met0_vl_x, met0_vl_y], callbacks=[early_stopper, check_pointer])

from tsm.eval_metrics import root_mean_squared_log_error

from keras.models import load_model
from keras.utils import CustomObjectScope
from keras.initializers import glorot_uniform

with CustomObjectScope({'GlorotUniform': glorot_uniform()}):
    model = load_model('models/DNN_MT0_261019T1415.json')

#dnn_tr_hat = (model.predict(met0_tr_x))
#print('Training NRMLSE:', root_mean_squared_log_error(dnn_tr_hat, met0_tr_y))

dnn_vl_hat = (model.predict(met0_vl_x))
vl_nrmlse = root_mean_squared_log_error(dnn_vl_hat, met0_vl_y)
print('Validation NRMLSE:', vl_nrmlse)
