==> nbs_arxiv/05-preprocessing-test.ipynb <==
%load_ext autoreload
%autoreload 2

import pandas as pd

## Test data

test_data = pd.read_pickle('kaggle/test_data.pkl')

import pandas as pd
from tsm.data_utils import time_processing, encode_categories, ordinal2wave
from tsm.data_selector import get_series_past_k_lags

test_data = time_processing(test_data, 'timestamp')

test_data = ordinal2wave('dt_m', test_data)
test_data = ordinal2wave('dt_w', test_data)
test_data = ordinal2wave('dt_d', test_data)
test_data = ordinal2wave('dt_hour', test_data)
test_data = ordinal2wave('dt_day_week', test_data)
test_data = ordinal2wave('dt_day_month', test_data)
test_data = ordinal2wave('dt_week_month', test_data)

Save here for the moment

%whos

test_data.to_pickle('kaggle/test_data.pkl')

test_data.head()

## To encode categories we need to have saved encoder at training time

from sklearn.externals import joblib

import numpy as np

encoder = joblib.load('models/LabelEconder_train_data.pkl')

test_data['primary_use'] = encoder.transform(test_data['primary_use']).astype(np.uint8)

test_data.head()

k_air = get_series_past_k_lags(test_data['air_temperature'], k=[1, 2, 3, 6]).fillna(99).reset_index(drop=True)
k_dew = get_series_past_k_lags(test_data['dew_temperature'], k=[1, 2, 3, 6]).fillna(99).reset_index(drop=True)
test_data = pd.concat([test_data.reset_index(drop=True), k_air, k_dew], axis=1, sort=False)

%whos

del k_dew

k_air_e_001 = test_data.air_temperature.ewm(alpha=0.01).mean().reset_index(drop=True)
k_air_e_001.name = 'k_air_e_001'

test_data = pd.concat([test_data.reset_index(drop=True), k_air_e_001], axis=1, sort=False)

del k_air_e_001

k_air_e_01 = test_data.air_temperature.ewm(alpha=0.1).mean().reset_index(drop=True)
k_air_e_01.name = 'k_air_e_01'

test_data = pd.concat([test_data.reset_index(drop=True), k_air_e_01], axis=1, sort=False)

del k_air_e_01

k_air_e_025 = test_data.air_temperature.ewm(alpha=0.25).mean().reset_index(drop=True)
k_air_e_025.name = 'k_air_e_025'

test_data = pd.concat([test_data.reset_index(drop=True), k_air_e_025], axis=1, sort=False)

del k_air_e_025

test_data.to_pickle('kaggle/proc_test_data.pkl')
