==> nbs_arxiv/06-making-predictions.ipynb <==
import pandas as pd

test_data = pd.read_pickle('/home/Ric/TSM/kaggle/test_data_meter_2.pkl')

test_data.dtypes

from sklearn.externals import joblib

mt0_mod = joblib.load('models/mt0_gbm_regress.pkl')
mt1_mod = joblib.load('models/mt1_gbm_regress.pkl')
mt2_mod = joblib.load('models/mt2_gbm_regress.pkl')
mt3_mod = joblib.load('models/mt3_gbm_regress.pkl')

models = {
    0: mt0_mod, 1: mt1_mod, 2: mt2_mod, 3: mt3_mod
}

from tqdm.auto import tqdm
tqdm.pandas()

feats = [x for x in list(test_data) if x not in ['row_id', 'meter', 'timestamp']]

test_data['yhat'] = test_data.progress_apply(lambda r: models[r.meter].predict(r[feats]), axis=1)

train = pd.read_pickle('/home/Ric/TSM/kaggle/train_data.pkl')

train.dtypes
